Refactored strip pattern
